+++ pkg/mutator/common_test.go
@@ -1,0 +1,10 @@
+package mutator
+
+import (
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+type GenericObject struct {
+	metav1.TypeMeta
+	metav1.ObjectMeta
+}